vectordb_PO_chatbot/requirements.txt
streamlit>=1.30.0
openai>=1.12.0
faiss-cpu>=1.7.4
pypdf>=4.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0